Move copying logic to helper function 'copyToClipboard'
Copy and download is implemented on the list and edit pages
