+++ src/helpers/downloadObjectAsJson.js
@@ -1,0 +1,11 @@
+const downloadObjectAsJson = (exportObj, exportName) => {
+    const dataStr = 'data:text/json;charset=utf-8,' + encodeURIComponent(JSON.stringify(exportObj))
+    let downloadAnchorNode = document.createElement('a')
+
+    downloadAnchorNode.setAttribute('href', dataStr)
+    downloadAnchorNode.setAttribute('download', exportName + '.json')
+    downloadAnchorNode.click()
+    downloadAnchorNode.remove()
+}
+
+export default downloadObjectAsJson
