+++ src/modules/modals/JSONmodal/JSONmodal.js
@@ -1,12 +1,13 @@
 import React, { PureComponent } from 'react'
 import { bool, object } from 'prop-types'
-import R from 'ramda'
 
 import Modal from '../../../components/Modal/Modal'
 import Button from '../../../components/Button/Button'
 import JSONeditor from '../../../components/JSONeditor/JSONeditor'
 
 import adjustToJSONeditor from '../../../helpers/adjustToJSONeditor'
+import copyToClipboard from '../../../helpers/copyToClipboard'
+import downloadObjectAsJson from '../../../helpers/downloadObjectAsJson'
 
 const propTypes = {
   message: object.isRequired,
@@ -38,12 +39,7 @@ class JSONmodal extends PureComponent {
   }
 
   handleCopyToClipBoard = () => {
-    const textarea = R.compose(
-      setValue(this.props.message),
-      createElement
-    )('textarea')
-
-    executeCopy(textarea)
+    copyToClipboard(this.props.message)
   }
 
   handleDownloadJSONfile = () => downloadObjectAsJson(this.props.message, this.props.message.name)
@@ -98,32 +94,3 @@ JSONmodal.propTypes = propTypes
 JSONmodal.defaultProps = defaultProps
 
 export default JSONmodal
-
-function executeCopy (textarea) {
-  document.body.appendChild(textarea)
-  textarea.select()
-  document.execCommand('Copy')
-  document.body.removeChild(textarea)
-}
-
-function setValue (value) {
-  return (element) => {
-    element.value = JSON.stringify(value)
-
-    return element
-  }
-}
-
-function createElement (name) {
-  return document.createElement(name)
-}
-
-function downloadObjectAsJson (exportObj, exportName) {
-  const dataStr = 'data:text/json;charset=utf-8,' + encodeURIComponent(JSON.stringify(exportObj))
-  let downloadAnchorNode = document.createElement('a')
-
-  downloadAnchorNode.setAttribute('href', dataStr)
-  downloadAnchorNode.setAttribute('download', exportName + '.json')
-  downloadAnchorNode.click()
-  downloadAnchorNode.remove()
-}
